Refactor source/header and add main.c (DOES NOT WORK YET)
